pyproject.toml
[project]
name = "kaizo"
version = "1.5.4"
description = "declarative YAML-based configuration parser"
authors = [{ name = "Mohammad Ghazanfari", email = "mgh.5225@gmail.com" }]
readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0.3", "typing-extensions>=4.15.0"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/NaughtFound/kaizo"
Source = "https://github.com/NaughtFound/kaizo"
Issues = "https://github.com/NaughtFound/kaizo/issues"
Changelog = "https://github.com/NaughtFound/kaizo/blob/main/CHANGELOG.md"
Documentation = "https://NaughtFound.github.io/kaizo"

[project.optional-dependencies]
hf = ["kaizo-hf"]

[dependency-groups]
test = ["pytest>=9.0.1"]
docs = ["Sphinx<=8.3.0", "sphinx-rtd-theme>=3.0.2"]

[build-system]
requires = ["setuptools >= 80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["kaizo*"]
